# Conflicts:
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/service/ReplenishmentSuggestionService.java
#	erp-server/erp-server-mrp/src/main/java/com/erp/server/mrp/service/impl/ReplenishmentSuggestionServiceImpl.java

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/listing/BodyBean.java
@@ -64,7 +64,7 @@ public class BodyBean {
     @JsonProperty("title")
     private String title;
     @JsonProperty("seller_id")
-    private int sellerId;
+    private long sellerId;
     @JsonProperty("category_id")
     private String categoryId;
     @JsonProperty("user_product_id")
